Add files via upload
Dakar, ‘a city of extremes’: beautiful highways and tiny bumpy roads, central business districts and precarious neighbourhoods, high fashion women and hip hop stars, bright coloured minibuses and so much more. Once a tiny settlement in the South of the Cap Vert peninsula, Dakar is now one of the most rapidly developing cities of Africa, with today a population over 2.4 million inhabitants. Over the last hundred years, Dakar has spread out from West to East over nearly 40 km, where it swallowed up a number of surrounding areas, now covering an area of 550 km2. At the same time, it is growing upwards and more densely populated. The historical lack of planning and the eviction of poor inhabitants from the city center has caused a rapid urbanization at the periphery, fed by the flow of migrants from various parts of Senegal or even from other countries in the region, which contributes to the development of informal housing in big parts of Dakar. This lack of planning is the cause of weak basic infrastructure systems, lagging public services and facilities, lack of water and energy disruptions is particularly found in the department of Pikine, which is considered ‘the double of Dakar’. Also, this area faces the emergency of new natural risks. Many informal and precarious settlements have historically been established in areas that were previously humid, but due to the drought of the 1970s to 1990s had dried up. Since 2000, the return of normal rainfall has caused severe flooding problems among the poorest population. Faced with this, the state, and with the help of various international organizations, has begun various programs of water management. With help of the World Bank, the lakes in the region will be connected to the Mbeubeuss lake, in order to manage the flood risks. Additionally, there is greater strategic planning, an increase in awareness campaigns and the construction of more advanced infrastructure. This urban planning, managing and organizing of the city causes various challenges. 

Addressing this challenges, this report aims to propose a series of improvement projects and programs within the area, reshaping the landscape, involving the whole community and placing sustainability at the center. This report is a collaborative work, based on an extensive research both academic and on-site.